// kyrvo-frontend/src/models/Agent.ts
export interface Agent {
  Id?: string;
  AgencyId?: string;
  FirstName?: string;
  LastName?: string;
  Phone?: string;
  Email?: string;
  Address1?: string;
  Address2?: string;
  City?: string;
  State?: string;
  PostalCode?: string;
  IsActive?: number;
  Remarks?: string;
  Meta?: string;
  CreatedOn?: number;
  UpdatedOn?: number;
}